% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homeRanger.R
\name{hrAsym}
\alias{hrAsym}
\title{Asymptote diagnostics for home range estimates}
\usage{
hrAsym(locs = NULL, type = "mcp", ID = NULL, levels = 90,
  percent = 95, a = NULL, unin = "m", unout = "ha", plotit = F,
  grid = 60, same4all = F, kern = "bivnorm", extent = 1,
  duplicates = "random", hlim = c(0.5, 1.5), arange = NULL,
  amount = "NULL", h = "href")
}
\arguments{
\item{locs}{is a SpatialPointsDataFrame containing coordinates(x,y) and 
a column of individual identifiers (ID).}

\item{type}{Defaults to 'mcp'. Options include 'mcp', 'kud' and 'LoCoH'}

\item{ID}{is a column from locs, and will be used for referencing individual-level diagnostics.}

\item{...}{further arguments, including optional home range arguments from adeHabitatHR}
}
\description{
The HRasym function allows users to examine whether and at what point the home range estimate
reaches an asymptote by iteratively adding relocations to a base subsample. 
The graph created by the HRasym function depicts the home range estimate for each iteration.
}
\examples{
hrDiag(locs, type = 'mcp', ID = locs$ID)

}
\keyword{home}
\keyword{range}
