% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_functions.R
\name{make_world}
\alias{make_world}
\title{Detection-explicit world building}
\usage{
make_world(rows, columns, det_dist = "random", min_det = 0, max_det = 1,
  shape1 = 10, shape2 = 10)
}
\arguments{
\item{rows}{is the number of rows in the simulated world}

\item{columns}{is the number of columns in the simulated world}

\item{det_dist}{is the statistical distribution from which cell-specific
detection probabilities will be drawn. Options include a random uniform,
'random', and beta distribution, 'beta'. Defaults to a random uniform between 0 and 1.}

\item{min_det}{is the minimum detection probability of any cell in the simulation}

\item{max_det}{is the maximum detection probability of any cell in the simulation}

\item{shape1}{is the first shape parameter of the beta distribution.}

\item{shape2}{is the second parameter of the beta distribution.}
}
\description{
make_world() builds a world according to your size and detection specifications
}
\examples{
make_world(rows = 10, columns = 10, det_dist = 'random', min_det = 0, max_det = 1)
}
\keyword{detection}
\keyword{probability}
