% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_functions.R
\name{move_critters}
\alias{move_critters}
\title{Simulate correlated random walks using the simulated population (pop_world) and the detection space(myworld)}
\usage{
move_critters(pop_world, myworld, world.type = "closed", Nsteps,
  homerange.type = "fixed", homerange.size, mu = 0, rho = 0)
}
\arguments{
\item{pop_world}{is the simulated population, created by populate_world()}

\item{myworld}{is a detection world, created by make_world()}

\item{world.type}{worlds may be 'open' such that orgs can leave the sample space, or 'closed' such that they cannot}

\item{Nsteps}{The number of timesteps in which the organisms can make movements}

\item{homerange.type}{Home ranges may be either 'fixed' with a defined home range for all organisms, or 'random' such that each individual draws a home range size in each iteration}

\item{homerange.size}{Home range size of individuals with a fixed home range}

\item{mu}{the mu parameter of the wrapped cauchy distribution determining step direction}

\item{rho}{the rho parameter of the wrapped cauchy distribution determining step direction}
}
\description{
Simulate correlated random walks using the simulated population (pop_world) and the detection space(myworld)
}
\examples{
move_critters(pop_world = pop_world, myworld = world, world.type = 'closed', Nsteps = 10, homerange.size = 10, mu = 0, rho = 0)
}
\keyword{correlated}
\keyword{random}
\keyword{walk}
