% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_functions.R
\name{populate_world}
\alias{populate_world}
\title{Populating simulated world with simulated organisims}
\usage{
populate_world(num_organisms, worldToPopulate, num_per_cell, replace = FALSE)
}
\arguments{
\item{num_organisms}{is the number of simulated organisms to init}

\item{worldToPopulate}{is a make_world() detection world}

\item{num_per_cell}{is the maximum number of oganisms which may share a cell}

\item{replace}{cells may be sampled with or without replacement}
}
\description{
The populate_world function initiates simulated organisms in a DiagnoseHR world
}
\examples{
populate_world(num_organisms = 3, worldToPopulate = world, num_per_cell = 1, replace = F)
}
\keyword{correlated}
\keyword{random}
\keyword{walk}
