% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_functions.R
\name{sample_world}
\alias{sample_world}
\title{Sample a movement record from a correlated random walk in a detection landscape}
\usage{
sample_world(world, walk, n.cells, sample.steps, replace.step = F,
  replace.world = T, seed = NULL)
}
\arguments{
\item{world}{The detection world to sample}

\item{walk}{The correlated random walk to sample}

\item{n.cells}{The number of cells to sample per timestep}

\item{sample.steps}{The number of time steps to sample}

\item{replace.step}{Whether to sample with replacement within a time step}

\item{replace.world}{Whether to sample with replacement within the entire simulation}

\item{seed}{An optional seed}
}
\description{
Sample a movement record from a correlated random walk in a detection landscape
}
\examples{
sample_world(world = world, walk = walk, sample.steps = seq(1,10,1), replace.step = F, replace.world = T, seed = NULL)
}
\keyword{detection}
\keyword{sampling,}
