% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_locs-data.R
\docType{data}
\name{sim_locs}
\alias{sim_locs}
\title{Sample simulated location data for home range diagnostics}
\format{a dataframe with 3 columns and 363 rows}
\usage{
data(sim_locs)
}
\description{
Data from a simulatio study assessing the interacting effects
of sampling schedule, detection, and home range estimation method.
locs has three variables: x, the x coordinate, y, the y coordinate,
and ID, an individual-level identifier.
}
\examples{
data(sim_locs)

}
\keyword{datasets}
